move Html.js out of helpers and into its own container
